--- client/src/pages/Home/HomeShop.css
@@ -1,24 +1,0 @@
-.images-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-.top-row {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
-
-.bottom-row {
-    width: 63%;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 5px;
-}
-
-.responsive-image {
-    border-radius: 5%;
-    width: 100%;
-    margin-right: 5px;
-}
